Add varstore and variable files to library.uefi
