+++ chipsec/library/uefi/uefi_common.py
@@ -23,80 +23,14 @@
 Common UEFI/EFI functionality including UEFI variables, Firmware Volumes, Secure Boot variables, S3 boot-script, UEFI tables, etc.
 """
 
-import os
 import struct
-import codecs
 from collections import namedtuple
-from collections.abc import Callable
-from typing import Dict, List, Tuple, Optional, Any
+from typing import Dict, List, Any
 from uuid import UUID
 
-from chipsec.library.file import read_file, write_file
-from chipsec.library.logger import logger
 from chipsec.library.defines import bytestostring
 
 # from chipsec.helper.oshelper import helper
-
-
-################################################################################################
-#
-# EFI Variable and Variable Store Defines
-#
-################################################################################################
-
-# UDK2010.SR1\MdeModulePkg\Include\Guid\VariableFormat.h
-#
-# Variable data start flag.
-#
-VARIABLE_DATA = 0x55aa
-VARIABLE_DATA_SIGNATURE = struct.pack('=H', VARIABLE_DATA)
-
-
-#
-# Variable Attributes
-#
-EFI_VARIABLE_NON_VOLATILE = 0x00000001  # Variable is non volatile
-EFI_VARIABLE_BOOTSERVICE_ACCESS = 0x00000002  # Variable is boot time accessible
-EFI_VARIABLE_RUNTIME_ACCESS = 0x00000004  # Variable is run-time accessible
-EFI_VARIABLE_HARDWARE_ERROR_RECORD = 0x00000008
-EFI_VARIABLE_AUTHENTICATED_WRITE_ACCESS = 0x00000010  # Variable is authenticated
-EFI_VARIABLE_TIME_BASED_AUTHENTICATED_WRITE_ACCESS = 0x00000020  # Variable is time based authenticated
-EFI_VARIABLE_APPEND_WRITE = 0x00000040  # Variable allows append
-EFI_VARIABLE_ENHANCED_AUTHENTICATED_ACCESS = 0x00000080
-UEFI23_1_AUTHENTICATED_VARIABLE_ATTRIBUTES = (EFI_VARIABLE_AUTHENTICATED_WRITE_ACCESS | EFI_VARIABLE_TIME_BASED_AUTHENTICATED_WRITE_ACCESS | EFI_VARIABLE_ENHANCED_AUTHENTICATED_ACCESS)
-
-
-def IS_VARIABLE_ATTRIBUTE(_c: int, _Mask: int) -> bool:
-    return ((_c & _Mask) != 0)
-
-
-def IS_EFI_VARIABLE_AUTHENTICATED(attr: int) -> bool:
-    return (IS_VARIABLE_ATTRIBUTE(attr, EFI_VARIABLE_AUTHENTICATED_WRITE_ACCESS) or 
-            IS_VARIABLE_ATTRIBUTE(attr, EFI_VARIABLE_TIME_BASED_AUTHENTICATED_WRITE_ACCESS) or
-            IS_VARIABLE_ATTRIBUTE(attr, EFI_VARIABLE_ENHANCED_AUTHENTICATED_ACCESS))
-
-
-MAX_VARIABLE_SIZE = 1024
-MAX_NVRAM_SIZE = 1024 * 1024
-
-
-def get_nvar_name(nvram: bytes, name_offset: int, isAscii: bool):
-    if isAscii:
-        nend = nvram.find(b'\x00', name_offset)
-        name = nvram[name_offset:nend].decode('latin1')
-        name_size = len(name) + 1
-        return (name, name_size)
-    else:
-        nend = nvram.find(b'\x00\x00', name_offset)
-        name = nvram[name_offset:nend].decode('utf-16le')
-        name_size = len(name) + 2
-        return (name, name_size)
-
-
-VARIABLE_SIGNATURE_VSS = VARIABLE_DATA_SIGNATURE
-
-
-VARIABLE_STORE_FV_GUID = UUID('FFF12B8D-7696-4C8B-A985-2747075B4F50')
 
 
 ################################################################################################
@@ -233,271 +167,6 @@ def get_3b_size(s_data: bytes) -> int:
     return (ord(s_str[0]) + (ord(s_str[1]) << 8) + (ord(s_str[2]) << 16))
 
 
-# #################################################################################################
-#
-# UEFI Variable (NVRAM) Parsing Functionality
-#
-# #################################################################################################
-
-SIGNATURE_LIST = "<16sIII"
-SIGNATURE_LIST_size = struct.calcsize(SIGNATURE_LIST)
-
-
-def parse_sha256(data):
-    return
-
-
-def parse_rsa2048(data):
-    return
-
-
-def parse_rsa2048_sha256(data):
-    return
-
-
-def parse_sha1(data):
-    return
-
-
-def parse_rsa2048_sha1(data):
-    return
-
-
-def parse_x509(data):
-    return
-
-
-def parse_sha224(data):
-    return
-
-
-def parse_sha384(data):
-    return
-
-
-def parse_sha512(data):
-    return
-
-
-def parse_x509_sha256(data):
-    return
-
-
-def parse_x509_sha384(data):
-    return
-
-
-def parse_x509_sha512(data):
-    return
-
-
-def parse_external(data):
-    return
-
-
-def parse_pkcs7(data):
-    return
-
-
-sig_types: Dict[str, Tuple[str, Callable, int, str]] = {
-    "C1C41626-504C-4092-ACA9-41F936934328": ("EFI_CERT_SHA256_GUID", parse_sha256, 0x30, "SHA256"),
-    "3C5766E8-269C-4E34-AA14-ED776E85B3B6": ("EFI_CERT_RSA2048_GUID", parse_rsa2048, 0x110, "RSA2048"),
-    "E2B36190-879B-4A3D-AD8D-F2E7BBA32784": ("EFI_CERT_RSA2048_SHA256_GUID", parse_rsa2048_sha256, 0x110, "RSA2048_SHA256"),
-    "826CA512-CF10-4AC9-B187-BE01496631BD": ("EFI_CERT_SHA1_GUID", parse_sha1, 0x24, "SHA1"),
-    "67F8444F-8743-48F1-A328-1EAAB8736080": ("EFI_CERT_RSA2048_SHA1_GUID", parse_rsa2048_sha1, 0x110, "RSA2048_SHA1"),
-    "A5C059A1-94E4-4AA7-87B5-AB155C2BF072": ("EFI_CERT_X509_GUID", parse_x509, 0, "X509"),
-    "0B6E5233-A65C-44C9-9407-D9AB83BFC8BD": ("EFI_CERT_SHA224_GUID", parse_sha224, 0x2c, "SHA224"),
-    "FF3E5307-9FD0-48C9-85F1-8AD56C701E01": ("EFI_CERT_SHA384_GUID", parse_sha384, 0x40, "SHA384"),
-    "093E0FAE-A6C4-4F50-9F1B-D41E2B89C19A": ("EFI_CERT_SHA512_GUID", parse_sha512, 0x50, "SHA512"),
-    "3bd2a492-96c0-4079-b420-fcf98ef103ed": ("EFI_CERT_X509_SHA256_GUID", parse_x509_sha256, 0x40, "X509_SHA256"),
-    "7076876e-80c2-4ee6-aad2-28b349a6865b": ("EFI_CERT_X509_SHA384_GUID", parse_x509_sha384, 0x50, "X509_SHA384"),
-    "446dbf63-2502-4cda-bcfa-2465d2b0fe9d": ("EFI_CERT_X509_SHA512_GUID", parse_x509_sha512, 0x60, "X509_SHA512"),
-    "452e8ced-dfff-4b8c-ae01-5118862e682c": ("EFI_CERT_EXTERNAL_MANAGEMENT_GUID", parse_external, 0x11, "EXTERNAL_MANAGEMENT"),
-    "4AAFD29D-68DF-49EE-8AA9-347D375665A7": ("EFI_CERT_TYPE_PKCS7_GUID", parse_pkcs7, 0, "PKCS7"),
-    }
-
-
-def parse_sb_db(db: bytes, decode_dir: str) -> List[bytes]:
-    entries = []
-    dof = 0
-    nsig = 0
-    db_size = len(db)
-    if 0 == db_size:
-        return entries
-
-    # some platforms have 0's in the beginnig, skip all 0 (no known SignatureType starts with 0x00):
-    while (dof + SIGNATURE_LIST_size) < db_size:
-        SignatureType0, SignatureListSize, SignatureHeaderSize, SignatureSize \
-            = struct.unpack(SIGNATURE_LIST, db[dof:dof + SIGNATURE_LIST_size])
-
-        # prevent infinite loop when parsing malformed var
-        if SignatureListSize == 0:
-            logger().log_bad("db parsing failed!")
-            return entries
-
-        # Determine the signature type
-        SignatureType = EFI_GUID_STR(SignatureType0)
-        sig_parse_f = None
-        sig_size = 0
-        if (SignatureType in sig_types.keys()):
-            sig_name, sig_parse_f, sig_size, short_name = sig_types[SignatureType]
-        else:
-            logger().log_bad(f'Unknown signature type {SignatureType}, skipping signature decode.')
-            dof += SignatureListSize
-            continue
-
-        # Extract signature data blobs
-        if (((sig_size > 0) and (sig_size == SignatureSize)) or ((sig_size == 0) and (SignatureSize >= 0x10))):
-            sof = 0
-            sig_list = db[dof + SIGNATURE_LIST_size + SignatureHeaderSize:dof + SignatureListSize]
-            sig_list_size = len(sig_list)
-            while ((sof + EFI_GUID_SIZE) < sig_list_size):
-                sig_data = sig_list[sof:sof + SignatureSize]
-                owner0 = struct.unpack(EFI_GUID_FMT, sig_data[:EFI_GUID_SIZE])[0]
-                owner = EFI_GUID_STR(owner0)
-                data = sig_data[EFI_GUID_SIZE:]
-                entries.append(data)
-                sig_file_name = f'{short_name}-{owner}-{nsig:02d}.bin'
-                sig_file_name = os.path.join(decode_dir, sig_file_name)
-                write_file(sig_file_name, data)
-                if (sig_parse_f is not None):
-                    sig_parse_f(data)
-                sof = sof + SignatureSize
-                nsig = nsig + 1
-        else:
-            err_str = f'Wrong SignatureSize for {SignatureType} type: 0x{SignatureSize:X}.'
-            if (sig_size > 0):
-                err_str = err_str + f' Must be 0x{sig_size:X}.'
-            else:
-                err_str = err_str + " Must be >= 0x10."
-            logger().log_error(err_str)
-            logger().log_error('Skipping signature decode for this list.')
-        dof = dof + SignatureListSize
-
-    return entries
-
-
-#
-#  "certdb" variable stores the signer's certificates for non PK/KEK/DB/DBX
-# variables with EFI_VARIABLE_TIME_BASED_AUTHENTICATED_WRITE_ACCESS|EFI_VARIABLE_NON_VOLATILE set.
-#  "certdbv" variable stores the signer's certificates for non PK/KEK/DB/DBX
-# variables with EFI_VARIABLE_TIME_BASED_AUTHENTICATED_WRITE_ACCESS set
-#
-# GUID: gEfiCertDbGuid
-#
-# We need maintain atomicity.
-#
-# Format:
-# +----------------------------+
-# | UINT32                     | <-- CertDbListSize, including this UINT32
-# +----------------------------+
-# | AUTH_CERT_DB_DATA          | <-- First CERT
-# +----------------------------+
-# | ........                   |
-# +----------------------------+
-# | AUTH_CERT_DB_DATA          | <-- Last CERT
-# +----------------------------+
-#
-# typedef struct {
-#   EFI_GUID    VendorGuid;
-#   UINT32      CertNodeSize;
-#   UINT32      NameSize;
-#   UINT32      CertDataSize;
-#   /// CHAR16  VariableName[NameSize];
-#   /// UINT8   CertData[CertDataSize];
-# } AUTH_CERT_DB_DATA;
-#
-AUTH_CERT_DB_LIST_HEAD = "<I"
-AUTH_CERT_DB_LIST_HEAD_size = struct.calcsize(AUTH_CERT_DB_LIST_HEAD)
-AUTH_CERT_DB_DATA = "<16sIII"
-AUTH_CERT_DB_DATA_size = struct.calcsize(AUTH_CERT_DB_DATA)
-
-
-def parse_auth_var(db: bytes, decode_dir: str) -> List[bytes]:
-    entries = []
-    dof = 0
-    nsig = 0
-    db_size = len(db)
-
-    # Verify that list makes sense
-    if db_size < AUTH_CERT_DB_LIST_HEAD_size:
-        logger().log_warning("Cert list empty.")
-        return entries
-    expected_size = struct.unpack(AUTH_CERT_DB_LIST_HEAD, db[dof:dof + AUTH_CERT_DB_LIST_HEAD_size])[0]
-    if db_size != expected_size:
-        logger().log_error("Expected size of cert list did not match actual size.")
-        return entries
-    dof += AUTH_CERT_DB_LIST_HEAD_size
-
-    # Loop through all the certs in the list.
-    while dof + AUTH_CERT_DB_DATA_size < db_size:
-        ven_guid0, cert_node_size, name_size, cert_data_size = struct.unpack(AUTH_CERT_DB_DATA, db[dof:dof + AUTH_CERT_DB_DATA_size])
-        vendor_guid = EFI_GUID_STR(ven_guid0)
-        name_size *= 2  # Name size is actually the number of CHAR16 in the name array
-        tof = dof + AUTH_CERT_DB_DATA_size
-        try:
-            var_name = codecs.decode(db[tof:tof + name_size], 'utf-16')
-        except UnicodeDecodeError:
-            logger().log_warning(f'Unable to decode {db[tof:tof + name_size]}')
-            var_name = "chipsec.library.exceptions!"
-        tof += name_size
-        sig_data = db[tof:tof + cert_data_size]
-        entries.append(sig_data)
-        sig_file_name = f'{vendor_guid}-{codecs.encode(var_name)}-{nsig:02X}.bin'
-        sig_file_name = os.path.join(decode_dir, sig_file_name)
-        write_file(sig_file_name, sig_data)
-        dof += cert_node_size
-        nsig += 1
-
-    return entries
-
-
-ESAL_SIG_SIZE = 256
-
-
-def parse_esal_var(db: bytes, decode_dir: str) -> List[bytes]:
-    entries = []
-    dof = 0
-    nsig = 0
-    db_size = len(db)
-
-    # Check to see how many signatures exist
-    if db_size < ESAL_SIG_SIZE:
-        logger().log('No signatures present.')
-        return entries
-
-    # Extract signatures
-    while dof + ESAL_SIG_SIZE <= db_size:
-        key_data = db[dof:dof + ESAL_SIG_SIZE]
-        entries.append(key_data)
-        key_file_name = os.path.join(decode_dir, f'AuthVarKeyDatabase-cert-{nsig:02X}.bin')
-        write_file(key_file_name, key_data)
-        dof += ESAL_SIG_SIZE
-        nsig += 1
-
-    return entries
-
-
-SECURE_BOOT_SIG_VAR = 1
-AUTH_SIG_VAR = 2
-ESAL_SIG_VAR = 3
-
-
-def parse_efivar_file(fname: str, var: Optional[bytes] = None, var_type: int = SECURE_BOOT_SIG_VAR) -> None:
-    logger().log(f'Processing certs in file: {fname}')
-    if not var:
-        var = read_file(fname)
-    var_path = fname + '.dir'
-    if not os.path.exists(var_path):
-        os.makedirs(var_path)
-    if var_type == SECURE_BOOT_SIG_VAR:
-        parse_sb_db(var, var_path)
-    elif var_type == AUTH_SIG_VAR:
-        parse_auth_var(var, var_path)
-    elif var_type == ESAL_SIG_VAR:
-        parse_esal_var(var, var_path)
-    else:
-        logger().log_warning(f'Unsupported variable type requested: {var_type}')
 
 
 # #################################################################################################
